requirements.txt
playwright
playwright-stealth
beautifulsoup4
APScheduler
langchain
langchain-community
langchain-text-splitters
langchain-core
langchain-groq
langchain-openai
langchain-google-genai
chromadb
langchain-chroma
pypdf
sentence-transformers
streamlit
python-dotenv
google-generativeai
pydantic
